+++ cmd/template/cluster/provider/azure.go
@@ -27,7 +27,7 @@ const (
 func WriteAzureTemplate(ctx context.Context, client k8sclient.Interface, out io.Writer, config ClusterCRsConfig) error {
 	var err error
 
-	if key.IsCAPZVersion(config.ReleaseVersion) {
+	if key.IsCAPIVersion(config.ReleaseVersion) {
 		err = WriteCAPZTemplate(ctx, client, out, config)
 		if err != nil {
 			return microerror.Mask(err)