+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Changed
+
+- Usa CAPI templates for all releases that start with `v20.0.0`, to include alpha and beta releases.
+
 ## [1.45.0] - 2021-10-26
 
 ### Added